Move Modals stylesheet in styles/stylesheet
