+++ fe1-web/components/DismissModal.tsx
@@ -1,44 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import {
-  StyleSheet, View, ViewStyle, Modal, TextStyle, Text,
+  View, Modal, Text,
 } from 'react-native';
 import STRINGS from 'res/strings';
-import { white } from 'styles/colors';
-import { Typography } from 'styles';
+import styles from 'styles/stylesheets/modal';
 import WideButtonView from './WideButtonView';
-
-const styles = StyleSheet.create({
-  modalView: {
-    backgroundColor: white,
-    borderRadius: 10,
-    borderWidth: 1,
-    margin: 'auto',
-    width: 550,
-  } as ViewStyle,
-  titleView: {
-    borderBottomWidth: 1,
-  } as ViewStyle,
-  modalTitle: {
-    ...Typography.important,
-    alignSelf: 'flex-start',
-    padding: 20,
-    paddingLeft: 10,
-  } as TextStyle,
-  modalDescription: {
-    ...Typography.base,
-    fontSize: 20,
-    alignSelf: 'flex-start',
-    textAlign: 'left',
-    padding: 20,
-    paddingLeft: 10,
-  } as TextStyle,
-  buttonView: {
-    alignSelf: 'center',
-    flexDirection: 'row',
-    paddingBottom: 20,
-  } as ViewStyle,
-});
 
 const DismissModal = (props: IPropTypes) => {
   const { visibility } = props;
